downgrade rust version and move vcpu.rs out
