+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ jobs:
     - uses: actions-rs/toolchain@v1
       with:
         profile: minimal
-        toolchain: nightly-2021-03-01
+        toolchain: nightly-2020-06-04
         override: true
         components: rustfmt, clippy
     - name: Check code format
@@ -30,7 +30,7 @@ jobs:
     - uses: actions-rs/toolchain@v1
       with:
         profile: minimal
-        toolchain: nightly-2021-03-01
+        toolchain: nightly-2020-06-04
         components: rust-src
         target: riscv64imac-unknown-none-elf
     - name: Build x86_64
@@ -60,7 +60,7 @@ jobs:
     - uses: actions-rs/toolchain@v1
       with:
         profile: minimal
-        toolchain: nightly-2021-03-01
+        toolchain: nightly-2020-06-04
         components: rust-src
     - name: Test
       uses: actions-rs/cargo@v1
